requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.2
fonttools==4.53.1
gunicorn==22.0.0
idna==3.7
kiwisolver==1.4.5
Markdown==3.6
matplotlib==3.9.1
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg-pool==3.2.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
